Implement config'd list of dynamically loaded modules
Instead of a fixed sequence of 4 module types, we now support an arbitrary
list of modules that are chained together in a pipeline.  Each module
is a dynamically loaded python module.  To implement a new module,
simply drop it in the sentry directory or other directory in the path.
Each module defines its own configuration validation schema.